[M144-LTS] Use 16-bit size for ResourceKeys
Internally, ResourceKey required the size to fit into a uint16_t so this
makes that explicit in the public API. It also changes how the size is
stored to instead record the num32DataCount directly and then convert to
bytes as needed, whereas previously it was requiring that the actual
byte count fit into a uint16_t. This gives a bit more head room.

Call sites to the ResourceKey builders are updated to now have the
responsibility of checking that their size can fit into a uint16_t. For
the most part, these were fixed or trivially small variable key sizes.
The two exceptions were Ganesh's style key (with dashes) and its
inherited key system for shapes with applied styles and path effects.
They now have reasonable limits to prevent the keys from growing bigger
than about 1kb.

Bug: b/495700484
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1199497
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
(cherry picked from commit 0566b2f5f0d1f218d9990eba838af16826d2e7e3)
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1206356
Commit-Queue: Thomas Smith <thomsmit@google.com>
Auto-Submit: Michael Ludwig <michaelludwig@google.com>
(cherry picked from commit bc591f8db342ee912fbb92aadde2a088e0ab8470)
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1215386

+++ src/gpu/graphite/mtl/MtlCaps.mm
@@ -953,7 +953,7 @@ const Caps::ColorTypeInfo* MtlCaps::getColorTypeInfo(
     return nullptr;
 }
 
-static constexpr int kMtlGraphicsPipelineKeyData32Count = 4;
+static constexpr uint16_t kMtlGraphicsPipelineKeyData32Count = 4;
 
 UniqueKey MtlCaps::makeGraphicsPipelineKey(const GraphicsPipelineDesc& pipelineDesc,
                                            const RenderPassDesc& renderPassDesc) const {
@@ -1241,7 +1241,7 @@ void MtlCaps::buildKeyForTexture(SkISize dimensions,
     SkASSERT(static_cast<uint32_t>(isFBOnly)    < (1u << 1));
 
     // We need two uint32_ts for dimensions, 2 for format, and 1 for the rest of the key;
-    static int kNum32DataCnt = 2 + 2 + 1;
+    static uint16_t kNum32DataCnt = 2 + 2 + 1;
 
     GraphiteResourceKey::Builder builder(key, type, kNum32DataCnt);
